% Generated by build script
\name{fec_committees}
\alias{fec_committees}
\title{Search FEC committees (PACs, party committees, campaign committees)}
\usage{
fec_committees()
}
\arguments{
\item{query}{Name search string}
\item{committee_type}{"P" (presidential), "H" (house), "S" (senate),}
\item{designation}{"A" (authorized by candidate), "B" (lobbyist/registrant),}
\item{api_key}{FEC API key}
\item{max_results}{Max results}
}
\value{tibble: committee_id, name, designation, committee_type, party, ...}
\description{Search FEC committees (PACs, party committees, campaign committees)}
