% Generated by build script
\name{fec_filings}
\alias{fec_filings}
\title{Search FEC filings}
\usage{
fec_filings()
}
\arguments{
\item{committee_id}{Committee ID filter}
\item{candidate_id}{Candidate ID filter}
\item{form_type}{Form type: "F1" (registration), "F2" (candidate auth),}
\item{min_receipt_date}{Start date}
\item{max_receipt_date}{End date}
\item{api_key}{FEC API key}
\item{max_results}{Max results}
}
\value{tibble of filing records}
\description{Search FEC filings}
