% Generated by build script
\name{fec_disbursements}
\alias{fec_disbursements}
\title{Committee disbursements (Schedule B)}
\usage{
fec_disbursements()
}
\arguments{
\item{committee_id}{Committee ID}
\item{recipient_name}{Recipient name search}
\item{min_date}{Start date}
\item{max_date}{End date}
\item{min_amount}{Minimum amount}
\item{two_year_transaction_period}{Election cycle}
\item{api_key}{FEC API key}
\item{max_results}{Max results}
}
\value{tibble of disbursement records}
\description{Committee disbursements (Schedule B)}
