% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fec.R
\name{fec_elections}
\alias{fec_elections}
\title{Election results and candidates by year/office/state/district}
\usage{
fec_elections(
  office,
  cycle,
  state = NULL,
  district = NULL,
  api_key = "DEMO_KEY"
)
}
\arguments{
\item{office}{"president", "senate", or "house"}

\item{cycle}{Election cycle year}

\item{state}{2-letter state code (required for senate/house)}

\item{district}{District number (required for house)}

\item{api_key}{FEC API key}
}
\value{
tibble of election results
}
\description{
Election results and candidates by year/office/state/district
}
